Rename and refactoring
