+++ package.json
@@ -2,9 +2,9 @@
   "name": "labs-authorization",
   "version": "0.0.1",
   "description": "Full stack authorization component",
-  "main": "integration/authorizationServer.js",
+  "main": "api/index.js",
   "scripts": {
-    "api:start": "node integration/index.js",
+    "api:start": "node api/index.js",
     "lint:service": "./node_modules/.bin/spacey-standard \"service/**/*.js\" \"test/**/*.js\"",
     "lint:component": "./component/node_modules/.bin/spacey-standard --parser babel-eslint \"component/src/**/*.js\" \"component/src/**/*.jsx\"",
     "lint": "npm run lint:service && npm run lint:component",
@@ -13,7 +13,7 @@
     "pg:stop": "docker-compose down",
     "pg:init": "[ ! -d ./logs ] && mkdir ./logs;./scripts/init/database/db_init.sh | tee ./logs/init-db.log",
     "pg:load-test-data": "./scripts/init/database/db_load.sh | tee ./logs/load-test-data.log",
-    "svc:start": "node integration/service.js",
+    "svc:start": "node service/index.js",
     "test:service": "tap \"test/**/*Test.js\" --coverage",
     "coverage:service": "tap \"test/**/*Test.js\" --cov --coverage-report=html",
     "test": "npm run test:service",
